==> notebooks/00_initial_exploration.ipynb <==
RAW_DATA_PATH = '../data/raw/survey_results_public.csv'

import pandas as pd
 


df = pd.read_csv(RAW_DATA_PATH)
df.head().T

df.shape

df.loc[:, 'EdLevel':'OpSys'].drop(columns=['Age1stCode']).columns

target_col = 'DevType'
possible_predictors  = df.loc[:, 'EdLevel':'OpSys'].drop(columns=['Age1stCode','CompFreq','CompTotal','Currency','LearnCode']).columns.to_list()

df_needed = df.loc[:, possible_predictors + [target_col]].copy()
df_needed.sample(4).T

df_needed.dtypes

# Invistigating questionable column types
cols = ['YearsCode','YearsCodePro']
for col in cols:
    print(f'Column {col} :')
    print(df_needed[col].unique())
    print(f'\n\n ','---'*20)

Needed preprocessing:
- Split Tools `;`
- Data type conversion: `YearsCode` , `YearsCodePro` from object to int16
    
    Issues: 'Less than 1 year' ,'More than 50 years' , nan

# serializing possible_predictors variable
import pickle


# serializing possible_predictors variable
with open('../data/variables/possible_predictors.pkl', 'wb') as f:
    pickle.dump(possible_predictors, f)
