==> notebooks/01_preprocessing.ipynb <==
# imports
import pandas as pd
import numpy as np
import pickle

RAW_DATA_PATH = '../data/raw/survey_results_public.csv'

# Constants:
PROCESSED_DATA_PATH = '../data/processed'
TARGET_COL = 'DevType'

with open('../data/variables/possible_predictors.pkl', 'rb') as f:
    PREDICTOR_COLS = pickle.load(f)

### Helper functions:

# functions:
def check_multiple_answers(df, col):
    return df[col].str.contains(';').any() == True

def split_answers(df, col):
    return df[col].str.split(';')      

df_raw = pd.read_csv(RAW_DATA_PATH).loc[:, PREDICTOR_COLS ]
df_raw.sample(3, random_state=42).T


# Preprocessing:

## Data Type conversion
- Data type conversion: `YearsCode` , `YearsCodePro` from object to int16
    
    Issues: 'Less than 1 year' ,'More than 50 years' , nan

df_processed = df_raw.copy()
# replacing 'Less than 1 year' with 0

df_processed.loc[df_processed['YearsCode'] == 'Less than 1 year', 'YearsCode'] = '0'
df_processed.loc[df_processed['YearsCodePro'] == 'Less than 1 year', 'YearsCodePro'] = '0'

# replacing 'More than 50 years' with 51
df_processed.loc[df_processed['YearsCode'] == 'More than 50 years', 'YearsCode'] = '51'
df_processed.loc[df_processed['YearsCodePro'] == 'More than 50 years', 'YearsCodePro'] = '51'


# converting to numeric
df_processed['YearsCode'] = df_processed['YearsCode'].astype(np.float32)
df_processed['YearsCodePro'] = df_processed['YearsCodePro'].astype(np.float32)

df_processed.dtypes

## Needed preprocessing:
-  Split multiple answers `;`

cat_cols = df_processed.select_dtypes(include=['object']).columns.tolist()
for col in cat_cols:
    print(f'Column {col} :')
    if check_multiple_answers(df_processed, col):
        print('Contains multiple answers')
        modified_series = split_answers(df_processed, col)
        mask_null = modified_series.isnull()
        modified_series.loc[mask_null] = modified_series.loc[mask_null].apply(lambda x: [])
        df_processed[col] = modified_series
        print('Splited')
    else:
        print('No multiple answers')
    print(f'\n','---'*5)
print('Done')


##### Verifying results

index = df_processed.sample(3).index
display(
    df_raw.loc[index, :] , 
    df_processed.loc[index, :]
)

### Export Data

df_processed.to_pickle(f'{PROCESSED_DATA_PATH}/1_preprocessed_df.pkl')
