==> notebooks/05_baseline.ipynb <==
import pandas as pd
import numpy as np
import mlflow
import pickle
import matplotlib.pyplot as plt
import plotly.express as px
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.multioutput import MultiOutputClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from mlflow.tracking import MlflowClient
from sklearn.metrics import precision_score

MLFLOW_TRACKING_URI = '../models/mlruns'
EXPERIMENT_NAME = 'skills_jobs_matching'
EXPORT_MODEL_PATH = '../models/baseline_model.pkl'
CLEANED_DF_PATH = '../data/processed/1_cleaned_df.pkl'
SAMPLED_DF_PATH = '../data/processed/1_sampled_df.pkl'


import os

os.makedirs(MLFLOW_TRACKING_URI, exist_ok=True)



df = pd.read_pickle(CLEANED_DF_PATH)

skills_columns = df.iloc[:,:24].columns.to_list()
jobs_columns = df.iloc[:,24:].columns.to_list()
display(
    skills_columns, 
    jobs_columns
)

X = df.loc[:,skills_columns]
y = df.loc[:,jobs_columns]

### Handling Imbalance

https://www.kaggle.com/code/rafjaa/resampling-strategies-for-imbalanced-datasets

y.sum().sort_values(ascending=False)

samples_per_class = 1200
resampled_roles = []
for role in y.columns:
    role_df = df[df[role] == 1].copy()
    if len(role_df) < samples_per_class:
        # upsample the minority class
        role_df = role_df.sample(samples_per_class, replace=True, random_state=42)
    else:
        # downsample the majority class
        role_df = role_df.sample(samples_per_class, random_state=42)
    resampled_roles.append(role_df)


resampled_df = pd.concat(resampled_roles, axis=0)
resampled_df.loc[:,y.columns].sum().sort_values(ascending=False)

# saving resampled_df as pickle file
resampled_df.to_pickle(SAMPLED_DF_PATH)

---
# Training Baseline model: logistics Regression

#### 1.Splitting data

X_resampled = resampled_df.loc[:,X.columns]
y_resampled = resampled_df.loc[:,y.columns]


X_train, X_test, y_train, y_test = train_test_split(X_resampled, y_resampled , test_size=0.3, random_state=42)


#### 2. Intializing MLflow

mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)



lg_model = Pipeline(steps = [
    ('scaler', StandardScaler()),
    ('classifier', MultiOutputClassifier(LogisticRegression()))
]
)
lg_model.fit(X_train, y_train)


# calculate precision, recall, and f1-score
from sklearn.metrics import classification_report
y_train_pred = lg_model.predict(X_train)
training_report = str(classification_report(y_train, y_train_pred, target_names=y_train.columns, zero_division=0)).split('\n')
training_report

y_test_pred = lg_model.predict(X_test)
test_report = str(classification_report(y_test, y_pred, target_names=y_train.columns, zero_division=0)).split('\n')
test_report

##### Export model

with open(EXPORT_MODEL_PATH, 'wb') as handle:
    pickle.dump(lg_model, handle)


### Track with Mlflow


# Initialize client and experiment
client = MlflowClient()
mlflow.set_experiment(EXPERIMENT_NAME)
exp = client.get_experiment_by_name(EXPERIMENT_NAME)


LOG_PATH = '../models/logs'

# make directory LOG_PATH
os.makedirs(LOG_PATH, exist_ok=True)

LOG_DATA_PKL = 'data.pkl'
LOG_MODEL_PKL = 'model.pkl'
LOG_METRICS_PKL = 'metrics.pkl'

data_details = {
    'data_path': CLEANED_DF_PATH,
    'training_indices': X_train.index.tolist(),
    'teseting_indices': X_test.index.tolist(),
    'features_names': X.columns.tolist(),
    'target_names': y.columns.tolist(),
}
with open(os.path.join(LOG_PATH, LOG_DATA_PKL), 'wb') as lg_file:
    pickle.dump(data_details, lg_file)

model = {
    'description': "Baseline model , Regression Model",
    'model_object': lg_model,
    'model_detailes': str(lg_model)
}
with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), 'wb') as lg_file:
    pickle.dump(model, lg_file)

metrics_details = {
    'training_details': training_report,
    'testing_details': test_report
}
with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), 'wb') as lg_file:
    pickle.dump(metrics_details, lg_file)

# Start a new run and track 
with mlflow.start_run(experiment_id=exp.experiment_id):
    mlflow.log_artifact(EXPORT_MODEL_PATH)   
    mlflow.log_metric("precision",precision_score(y_test, y_test_pred, average='macro', zero_division=0) ) 


runs = mlflow.search_runs([exp.experiment_id])


runs

best_run = runs.sort_values('metrics.precision').iloc[0]
best_run

# # Load model
# artifact_path = best_run["artifact_uri"].replace("file://", "")
# model_pkl = os.path.join(artifact_path, MODEL_PATH)
# with open(model_pkl, "rb") as f:
#     model = pickle.load(f)
    
# model
