==> notebooks/07_xgboost.ipynb <==
import pandas as pd
import numpy as np
import mlflow
import pickle
import plotly.express as px
from sklearn.model_selection import train_test_split
from sklearn.multioutput import MultiOutputClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from mlflow.tracking import MlflowClient
from sklearn.metrics import precision_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import RobustScaler
MLFLOW_TRACKING_URI = '../models/mlruns'
EXPERIMENT_NAME = 'skills_jobs_matching'
EXPORT_MODEL_PATH = '../models/rf_model.pkl'
DF_PATH = '../data/processed/1_sampled_df.pkl'
LOG_PATH = '../models/logs'
LOG_DATA_PKL    =  "xgb_data.pkl"
LOG_MODEL_PKL   =  "xgb_model.pkl"
LOG_METRICS_PKL =  "xgb_metrics.pkl"

df = pd.read_pickle(DF_PATH)
X = df.iloc[:,:25]
y = df.iloc[:,25:]

X_train, X_test, y_train, y_test = train_test_split(X,y , test_size=0.3, random_state=42)

mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)


### Random Forest Pipelining and Training

From EDA we noticed colinearity between features, in this pipeline we handle via PCA. But we first need to scale the features. 

from sklearn.decomposition import PCA
from xgboost import XGBClassifier

xgb_pipeline = Pipeline(steps=[
    ('scaler', RobustScaler()),
    ('pca', PCA(n_components=0.95)), 
    ('classifier', XGBClassifier(
        n_estimators=500,
        learning_rate=0.01,
        n_jobs=-1, 
        random_state=42
        
    ))
])
xgb_pipeline.fit(X_train, y_train)

from sklearn.metrics import classification_report
y_train_pred = xgb_pipeline.predict(X_train)
training_report = str(classification_report(y_train, y_train_pred, target_names=y_train.columns, zero_division=0)).split('\n')
training_report

from sklearn.metrics import classification_report
y_test_pred = xgb_pipeline.predict(X_test)
testing_report = str(classification_report(y_test, y_test_pred, target_names=y_train.columns, zero_division=0)).split('\n')
testing_report

We observe a strong sign of overfitting. Let's fine tune our model.

### Fine Tunining

from sklearn.model_selection import RandomizedSearchCV
from scipy.stats import uniform, randint


from sklearn.model_selection import RandomizedSearchCV
from scipy.stats import uniform, randint

param_dist = {
    'classifier__n_estimators': randint(200, 500),
    'classifier__learning_rate': uniform(0.01, 0.3),  # Range: [0.01, 0.31]
    'classifier__max_depth': randint(3, 10),
    'classifier__subsample': uniform(0.6, 0.4),       # Range: [0.6, 1.0]
    'classifier__reg_alpha': uniform(0.0, 1.0),       # L1 regularization
    'pca__n_components': [0.80, 0.90],
}

# Set up random search
random_search = RandomizedSearchCV(
    estimator=xgb_pipeline,
    param_distributions=param_dist,
    n_iter=10,                  # Try 30 random combinations
    scoring='precision',         # or 'f1', 'roc_auc' depending on task
    cv=3,
    verbose=1,
    random_state=42,
    n_jobs=-1
)

# Fit the search
random_search.fit(X_train, y_train)


random_search.best_params_

xgb_pipeline = random_search.best_estimator_
xgb_pipeline

y_train_pred = xgb_pipeline.predict(X_train)
training_report = str(classification_report(y_train, y_train_pred, target_names=y_train.columns, zero_division=0)).split('\n')
training_report

y_test_pred = xgb_pipeline.predict(X_test)
testing_report = str(classification_report(y_test, y_test_pred, target_names=y_train.columns, zero_division=0)).split('\n')
testing_report

#### Export Model

with open(EXPORT_MODEL_PATH, 'wb') as handle:
    pickle.dump(xgb_pipeline, handle)


### Track with Mlflow


# Initialize client and experiment
client = MlflowClient()
mlflow.set_experiment(EXPERIMENT_NAME)
exp = client.get_experiment_by_name(EXPERIMENT_NAME)


import os
data_details = {
    'data_path': DF_PATH,
    'training_indices': X_train.index.tolist(),
    'teseting_indices': X_test.index.tolist(),
    'features_names': X.columns.tolist(),
    'target_names': y.columns.tolist(),
}
with open(os.path.join(LOG_PATH, LOG_DATA_PKL), 'wb') as lg_file:
    pickle.dump(data_details, lg_file)

model = {
    'description': "Xgboost",
    'model_object': xgb_pipeline,
    'model_detailes': str(xgb_pipeline),
}
with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), 'wb') as lg_file:
    pickle.dump(model, lg_file)

metrics_details = {
    'training_details': training_report,
    'testing_details': testing_report
}
with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), 'wb') as lg_file:
    pickle.dump(metrics_details, lg_file)

# Start a new run and track 
with mlflow.start_run(experiment_id=exp.experiment_id):
    mlflow.log_artifact(EXPORT_MODEL_PATH)   
    mlflow.log_metric("precision",precision_score(y_test, y_test_pred, average='macro', zero_division=0) ) 


runs = mlflow.search_runs([exp.experiment_id])
runs
