==> notebooks/09_prediction.ipynb <==
import mlflow
from mlflow.tracking import MlflowClient
import pandas as pd
import os
import pickle


MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_EXPERIMENT_NAME = 'skills_jobs_matching'
MLFLOW_RUN_ID = '812636a81de04341b6d3ff99dd1037a6'
LOG_PATH = '../models/logs'
LOG_DATA_PKL    =  "rf_data.pkl"
LOG_MODEL_PKL   =  "rf_model.pkl"
LOG_METRICS_PKL =  "rf_metrics.pkl"

# Loading Model and Data

MLflow Settup


mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()
run = mlflow.get_run(MLFLOW_RUN_ID)
artifact_path = run.info.artifact_uri

artifact_path

model_path = '../models/mlruns/344207504652916142/812636a81de04341b6d3ff99dd1037a6/artifacts/temp/rf_model.pkl'
with open(model_path, "rb") as f:
    model = pickle.load(f)
model

data_path = '../models/mlruns/344207504652916142/812636a81de04341b6d3ff99dd1037a6/artifacts/temp/rf_data.pkl'
with open(data_path, "rb") as f:
    data = pickle.load(f)
data.keys()

classifier = model['model_object']
features_names = pd.Series(data['features_names'])
targets_names = pd.Series(data['target_names'])

pd.Series(data['features_names'])

targets_names

---
## Load Skills Clusters

import json
SKILLS_CLUSTERS_PATH = '../data/processed/skills_clusters.json'
with open(SKILLS_CLUSTERS_PATH, 'r') as f:
    skills_clusters = json.load(f)
skills_clusters = pd.Series(skills_clusters)

skills_clusters_df = pd.DataFrame([(cluster, skill) for cluster, skills in skills_clusters.items() for skill in skills], columns=['cluster_id', 'skill'])
skills_clusters_df

---
### Predict Sample entry


def get_skills_clusters(sample_skills, skills_clusters_df, features_names):
    """
    Get the skills clusters.
    
    Args:
        sample_skills: List or Series of skills to cluster
        skills_clusters_df: DataFrame mapping skills to clusters
        features_names: List of all possible cluster IDs
        
    Returns:
        List of cluster frequencies for the given skills
    """
    # Filter skills that exist in our clusters dataframe
    valid_skills = skills_clusters_df[skills_clusters_df['skill'].isin(sample_skills)]
    
    # Count skills per cluster
    clusters_freq = valid_skills['cluster_id'].value_counts().reindex(features_names, fill_value=0)
    
    return clusters_freq.tolist()

def make_predictions(input_data, model, features_names, targets_names, skills_clusters_df):
    # Prepare the input data
    clusters = get_skills_clusters(input_data, skills_clusters_df, features_names)
    input_df = pd.DataFrame([clusters], columns=features_names)
    
    # Make predictions
    predictions = model.predict_proba(input_df)
    
    positive_probs= [prob[0][1] for prob in predictions]
    
    return pd.Series(positive_probs, 
          index=targets_names).sort_values(ascending=False)


    



sample_skills = pd.Series(['Scala','Python','Keras','Python'])

if not all(skill in skills_clusters_df['skill'].values for skill in sample_skills):
    print('Invalid skills provided.')
else:
    predictions = make_predictions(
        input_data=sample_skills,
        model=classifier,
        features_names=features_names,
        targets_names=targets_names,
        skills_clusters_df=skills_clusters_df
    )
    print("Predicted labels:\n", predictions)

import sys
import os

# Step 1: Add the scripts directory to sys.path
scripts_path = os.path.abspath(os.path.join(os.getcwd(), '..', 'scripts'))
if scripts_path not in sys.path:
    sys.path.append(scripts_path)

# Step 2: Import your class
from predictor import predict_job_probabilities



predict_job_probabilities(['Scala','Python','Keras','Python'])
